test-requirements.txt
vcrpy>=4.2.1
pytest>=7.2.0
coverage>=6.5.0
black>=22.10.0
pytest-mock>=3.10.0